LayoutTests/media/modern-media-controls/tracks-support/off-text-track-expected.txt

Ensure we have the Off text track selected by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became different from 0
Tapping tracks button...
PASS checkedItems.length is 1
PASS checkedItems[0].title is "Off"
PASS successfullyParsed is true

TEST COMPLETE
